third_party/blink/web_tests/http/tests/security/mixedContent/insecure-fetch-on-dedicated-worker-blocked.https-expected.txt

CONSOLE ERROR: Mixed Content: The page at 'https://127.0.0.1:8443/security/mixedContent/insecure-fetch-on-dedicated-worker-blocked.https.html' attempted to load subresource at 'https://127.0.0.1:8443/resources/testharness.js' over broken HTTPS.
CONSOLE ERROR: Mixed Content: The page at 'https://127.0.0.1:8443/security/mixedContent/insecure-fetch-on-dedicated-worker-blocked.https.html' attempted to load subresource at 'https://127.0.0.1:8443/resources/testharnessreport.js' over broken HTTPS.
CONSOLE ERROR: Mixed Content: The page at 'https://127.0.0.1:8443/security/mixedContent/insecure-fetch-on-dedicated-worker-blocked.https.html' attempted to load subresource at 'https://127.0.0.1:8443/security/mixedContent/resources/insecure-fetch-in-dedicated-worker.js' over broken HTTPS.
CONSOLE ERROR: line 1: Mixed Content: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/insecure-fetch-in-dedicated-worker.js' was loaded over HTTPS, but requested an insecure resource 'http://example.test:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=*'. This request has been blocked; the content must be served over HTTPS.
This is a testharness.js-based test.
PASS Insecure resource loads inside Worker failes when 'AllowRunningInsecureContent' is not set.
Harness: the test ran to completion.
